+++ netket/operator/_prod/discrete_operator.py
@@ -1,4 +1,4 @@
-# Copyright 2025 The NetKet Authors - All rights reserved.
+# Copyright 2021 The NetKet Authors - All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,38 +12,29 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import TYPE_CHECKING
-
 import numpy as np
 
 from netket.operator._discrete_operator import DiscreteOperator
-from netket.operator._sum.base import SumOperator
 
-if TYPE_CHECKING:
-    from netket.operator._sum.discrete_jax_operator import SumDiscreteJaxOperator
+from netket.operator._prod.base import ProductOperator
 
 
-class SumDiscreteOperator(SumOperator, DiscreteOperator):
-    def __init__(self, *operators, coefficients=1.0, dtype=None):
+class ProductDiscreteOperator(ProductOperator, DiscreteOperator):
+    def __init__(self, *operators, coefficient=1.0, dtype=None):
         if not all(isinstance(op, DiscreteOperator) for op in operators):
             raise TypeError(
-                "Arguments to SumDiscreteOperator must all be "
+                "Arguments to ProductDiscreteOperator must all be "
                 "subtypes of DiscreteOperator. However the types are:\n\n"
                 f"{list(type(op) for op in operators)}\n"
             )
         self._initialized = False
         super().__init__(
-            operators, operators[0].hilbert, coefficients=coefficients, dtype=dtype
+            operators, operators[0].hilbert, coefficient=coefficient, dtype=dtype
         )
 
     def _setup(self, force: bool = False):
         if not self._initialized:
             self._initialized = True
-
-    @property
-    def max_conn_size(self) -> int:
-        """The maximum number of non zero ⟨x|O|x'⟩ for every x."""
-        return sum(op.max_conn_size for op in self.operators)
 
     def get_conn_flattened(
         self,
@@ -53,14 +44,16 @@ class SumDiscreteOperator(SumOperator, DiscreteOperator):
     ) -> tuple[np.ndarray, np.ndarray]:
         raise NotImplementedError
 
-    def to_jax_operator(self) -> "SumDiscreteJaxOperator":
+    def to_jax_operator(self) -> "ProductDiscreteJaxOperator":  # noqa: F821
         """
         Returns the standard (numba) version of this operator, which is an
         instance of {class}`nk.operator.Ising`.
         """
-        from netket.operator._sum.discrete_jax_operator import SumDiscreteJaxOperator
+        from netket.operator._prod.discrete_jax_operator import (
+            ProductDiscreteJaxOperator,
+        )
 
-        ops_numba = tuple(op.to_jax_operator() for op in self.operators)
-        return SumDiscreteJaxOperator(
-            *ops_numba, coefficients=self.coefficients, dtype=self.dtype
+        ops_jax = tuple(op.to_jax_operator() for op in self.operators)
+        return ProductDiscreteJaxOperator(
+            *ops_jax, coefficient=self.coefficient, dtype=self.dtype
         )