+++ netket/utils/model_frameworks/nnx.py
@@ -83,8 +83,12 @@ class NNXWrapper:
         if hasattr(_get_graphdef_type(self.graphdef), name):
             return partial(self.apply, method=name)
         raise AttributeError(
-            f"'{type(self).__name__}' (and the wrapped '{self.graphdef.type}') object has no attribute '{name}'"
+            f"{_get_graphdef_type(self.static_module)} (wrapped into a '{type(self).__name__}')"
+            f"has no attribute '{name}'"
         )
+
+    def __hash__(self):
+        return hash(("NNXWrapper", self.graphdef))
 
     def __repr__(self):
         typ = _get_graphdef_type(self.graphdef)